Port range

+++ IM/connectors/OpenNebula.py
@@ -545,10 +545,10 @@ class OpenNebulaCloudConnector(CloudConnector):
 
          Returns: bool, True if there are at least one lease free or False otherwise
         """
-        start = long(''.join(["%02X" % long(i)
-                              for i in ar_range.IP_START.split('.')]), 16)
-        end = long(''.join(["%02X" % long(i)
-                            for i in ar_range.IP_END.split('.')]), 16)
+        start = int(''.join(["%02X" % int(i)
+                             for i in ar_range.IP_START.split('.')]), 16)
+        end = int(''.join(["%02X" % int(i)
+                           for i in ar_range.IP_END.split('.')]), 16)
         if end - start > int(total_leases):
             return True
         return False
